requirements.txt
openai
python-dotenv
pvporcupine
pyaudio
SpeechRecognition